Move the eligible intervening Bitcast between AR and DS ahead of AR to enable RS fusion.
The parsing codes for ReduceScatter fusion does not allow an intervening Bitcast between AR and DS. This cl detects such pattern and moves the Bitcast before AR and the Bitcast converts the original AR input to the target shape. This is done only when the Bitcast shape matches DS' shape, namely eligible for RS fusion.

PiperOrigin-RevId: 627481897