[xla:gpu] Fix a bug in pipelined-p2p-rewriter.
Previously, we call module schedule update each time when we transform a
while-body as well as at the end of the pass. It turns out that updating the
schedule for each while-body is not necessary and can also bring in changes
that aren't track and cause problem in keeping track of the instructions in
instruction_sequence. This CL removes the updating for each while-body.

Add a test.

PiperOrigin-RevId: 636190061