PR #13278: Fix _xla_send_recv_validation_attribute in loop-double-buffer-transformer
Imported from GitHub PR https://github.com/openxla/xla/pull/13278

This patch fixes the value of valid iteration for GPUs in the loop-double-buffer-transformer pass.

If the original iteration bounds for a particular GPU are {a,b} then the transformed attribute is:

  - If number of iterations is even:
    - For collective.permute.1: {floor((a+1)/2), floor(b/2)}
    - For collective.permute.2: {floor(a/2), max(0, floor((b-1)/2))}
  - If the number of iterations is odd: then the first iteration is peeled out
    - For collective.permute.1: {floor(a/2), max(0, floor((b-1)/2))}
    - For collective.permute.2: {max(0, floor((a-1)/2)), max(0, floor((b-2)/2)}

For collective.permute.peeled, only one instance of the collective is valid. It is valid for all
the devices that had iteration 0 as valid iteration earlier. So, if the attribute was
{{0,4},{0,5},{1,5},{1,6}} for the old instruction, then for the peeled instruction, the attribute
would be {{0,0},{0,0},{1,0},{1,0}} because it is valid for the first two devices, but invalid for
the rest.

More details for these are added in comments as docstrings.
Copybara import of the project:

--
050ea0994c31f2e9f87b8d1664190435578deee4 by Shraiysh Vaishay <svaishay@nvidia.com>:

Fix _xla_send_recv_validation_attribute in loop-double-buffer-transformer

This patch fixes the value of valid iteration for GPUs in the loop-double-buffer-transformer pass.

If the original iteration bounds for a particular GPU are {a,b} then the transformed attribute is:

  - If number of iterations is even:
    - For collective.permute.1: {floor((a+1)/2), floor(b/2)}
    - For collective.permute.2: {floor(a/2), max(0, floor((b-1)/2))}
  - If the number of iterations is odd: then the first iteration is peeled out
    - For collective.permute.1: {floor(a/2), max(0, floor((b-1)/2))}
    - For collective.permute.2: {max(0, floor((a-1)/2)), max(0, floor((b-2)/2)}

For collective.permute.peeled, only one instance of the collective is valid. It is valid for all
the devices that had iteration 0 as valid iteration earlier. So, if the attribute was
{{0,4},{0,5},{1,5},{1,6}} for the old instruction, then for the peeled instruction, the attribute
would be {{0,0},{0,0},{1,0},{1,0}} because it is valid for the first two devices, but invalid for
the rest.

More details for these are added in comments as docstrings.

Merging this change closes #13278

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/openxla/xla/pull/13278 from shraiysh:loop-double-buffer 050ea0994c31f2e9f87b8d1664190435578deee4
PiperOrigin-RevId: 642627932