Make NvPtxCompiler work without ptxas when libnvptxcompiler is enabled
Running XLA with `--xla_gpu_enable_libnvptxcompiler` makes all the PTX compilation go through libnvptxcompiler but there is still a code path
which is checking whether the version of ptxas is sufficient.

This change makes it check the version of libnvptxcompiler instead which
means XLA can truly run without ptxas.

PiperOrigin-RevId: 649044298