[XLA:MSA] Introduce WindowPrefetchedAllocation
WindowPrefetchedAllocation enables MSA to perform window prefetch. Previously, MSA can only prefetch a whole tensor, window prefetch allows MSA to prefetch only a window worth of data. WindowPrefetchAllocation allocates buffers in the alternate memory that hold windows of data.

WindowPrefetchedAllocations kick in only when a tensor was decided to be placed in the default memory. When the tensor was placed in the default memory, the MSA now will try to allocate a small buffer in the alternate memory using WindowPrefetchedAllocations. If the allocation succeeds, it will insert a pair of async WindowPrefetch operations, append a new operand to the use instruction and feed the output of the WindowPrefetch operation as the new operand. So the use instruction can use the buffer when being lowered.

If the allocation fails, window prefetch will not be used. Then this is the default behavior as today.

For simplicity, WindowPrefetch does not use sliced copy.

The insertion of WindowPrefetchedAllocation is gated under the `memory_space_assignment_options` field `enable_window_prefetch`, which is currently disabled, so this CL should be a no-op for now.

PiperOrigin-RevId: 650744700