[XLA:GPU][NFC] First commit splitting out `triton_fusion_emitter_test.cc` into several BUILD targets.
The file is for now split into two files, which still both require a device to
run:

1. `triton_fusion_emitter_device_test.cc`, which is intended to contain all
the tests for the new Triton emitters that require a device to run;
2. `triton_fusion_emitter_device_legacy_test.cc`, which is
`triton_fusion_emitter_device_test.cc`'s counterpart for the legacy GEMM
codegen path.

Eventually, the tests will be split further into `lowering` test files, which
will contain all the tests that do not require a device to run.

Also make sure to not run the optimization pipeline in tests for the new
Triton emitter---this is unnecessary and was a previous oversight.

PiperOrigin-RevId: 653207961