Add tests for PTX compilation in NVPTXCompiler
NVPTXCompiler (and the whole Compiler class hierarchy for that matter)
is a pretty monolithic component with very little test coverage.

The whole things needs a rearchitecturing to make it properly testable,
but for now I'm trying to strike a balance between not changing too many
things and allowing me to add NVJitLink support with decent coverage.

So this is adding a test harness for PTX compilation that supports
all the different compilation methods that we have.

NVJitlink support will come in a separate change.

PiperOrigin-RevId: 653228982