[XLA:FFI] Add API version to XLA_FFI_Api struct, and check at runtime.
This exposes the FFI API version as part of the API struct and adds a check at runtime to make sure that the major versions match. Eventually we will want to include logic to check the minor version like in Pjrt.

It would probably be a somewhat better user experience to error at registration time rather than call time, but we don't currently have enough metadata at registration time. It would be probably be possible to update the handler bundle provided when registering to make this work, but this was a much less invasive change, and since we already check the struct size at runtime here it seemed reasonable to include the version check in the same place.

I will update with benchmark info once I get a chance to run it.

PiperOrigin-RevId: 655663605