Enhance the method to handle sort in spmd partitioner.
cl/504334258 introduces an algorithm to partitioning the sort. If we partition the sorted dimension, we can use all-to-all to move the sharding tiles to a free-axis. A free axis satisfies
* It is not a sorted dimension.
* The dimension size is larger than 1.
* It is fully replicated, i.e., its tile size is 1.

We relax the criterion in this change.
* It is not a sorted dimension.
* The dimension size can be divided by the merged tile sizes.

The new criterion can match the following pattern.
```
shape: [4,32,32]
raw_sharding for all operands/results: [4,4,4]<=[64]
```
We can reshard (all-to-all) the operand to [4,16,1]<=[64] and then apply the sorting.
PiperOrigin-RevId: 693854434