[XLA:SchedulingAnnotations] Switch from the root readiness model to the scheduled successors model to launch an annotated scheduling group.
Root readiness does not cover the newly added test case where a producer-consumer computation pair is annotated with the same group id and the producer also feeds into some other op. We previously identified the consumer as the root and tried to schedule the annotation as soon as the consumer was ready (assuming the async collective is also ready). However, the non-annotated consumer of the producer might have yet to be scheduled.

In the new model, we make sure all (non-annotated) successors of the group are scheduled before we attempt to schedule it.

PiperOrigin-RevId: 730929730