[IFRT] Add short form for specifying platform_names for IFRT IR passes.
Some IFRT IR passes require a list of platform names to be given as an option. Currently, a platform names list requires an entry for each device, which makes  manually running the passes on modules with many devices tedious. This change introduces a short form for specifying platform names with the format  platform_name:number_of_occurrences. For example, tpu:2,cpu:3 is expanded to tpu,tpu,cpu,cpu,cpu.

PiperOrigin-RevId: 730967340