Make shape categories mutually exclusive.
Given the large number of existing bugs the desired change uncovers, this change is deliberately conservative: in some cases where the caller violates the precondition of a `Shape` method, we choose to return a default value instead of crashing (e.g. calling `dimensions()` on a non-array shape would return an empty span even though it's a programmer error). We will tightened the enforcement of the preconditions in future CLs.

PiperOrigin-RevId: 738606542