Add contracting split (split-K) selection logic to dynamic search space
First step towards the full search space. We figure out the contracting split
based on the problem size, and a few other properties of the search space (still hardcoded for now).

We also have an option for forcing a specific split, which is helpful for both
disabling autotuning that parameter, and will facilitate support for analytically setting it in the future.

PiperOrigin-RevId: 741479773