[PjRt-IFRT] Treat IFRT HloSharding with a single tile as a fully replicated sharding
`xla::ifrt::HloSharding` that has a single tile would have the shard buffer to
be the same as the global array both in the content and the shape. Thus, we
treat it as a fully replicated sharding in the context of IFRT APIs even though
`HloSharding` does not explicitly say full replication in the contxt of the XLA
`HloSharding`. This allows taking a runtime path for array creation specialized
for fully replicated sharding for a single-tile `HloSharding`.

PiperOrigin-RevId: 742864400