Removed `CoordinationServiceInterface`.
Previously, there was a pure virtual `CoordinationServiceInterface` interface
with a single subclass called `CoordinationServiceStandaloneImpl`. I think the
idea was that there would be multiple implementations of the
`CoordinationServiceInterface`, but there are not, and the extra layer of
indirection is a nuisance. This CL renames `CoordinationServiceStandaloneImpl`
to just `CoordinationService` and replaces `CoordinationServiceInterface` with
it.

PiperOrigin-RevId: 744787566