Move CUB sort kernels behind FFI
This means the cub_sort_thunk and the sort_rewriter
won't directly depend on the CUB implementation anymore.

Instead they go through the FFI registry.
This decouples CUDA and ROCm implementations of CUB from the
users of CUB and will also help with splitting the
compiler from the runtime.

This also removes the requirement for having a stub for SortRewriter.

PiperOrigin-RevId: 745027614