Add basic thunk deserialization logic
This adds the basic infra to deserialize a tree of ThunkProtos.

The deserializer is currently using stack-based recursion which is
easier to read then the alternatives and should be no issue for us
since Thunk trees are only 2-3 levels deep. If that changed in the future
we could easily change to some solution that doesn't use recursion.

For each Thunk type the following needs to be done:

1. The thunk type `XThunk` needs to get a `absl::StatusOr<std::unique_ptr<XThunk>> XThunk::FromProto(const XThunkProto&, ...)` static factory function
2. The global deserialization dispatch function `DeserializeThunkProto` needs to get a branch for the thunk type `XThunk`.

The change is doing that for the SequentialThunk

PiperOrigin-RevId: 755332119