[IFRT] Use SerDes versioning in most IFRT and PjRt-IFRT types
The current serialization format is anchored as `SerDesVersion::kV0Initial`.

All `ToProto()`s get a `version` parameter (defaults to
`SerDesVersion::current()`), and use it for internal format selection and
forward it to nested serialization.

All `Serialize()`s recognize `options` and get a version if provided (if
omitted, `SerDesVersion::current()` is assumed), and behave similarly to
`ToProto()`.

`Deserialize()`/`FromProto()` will explicitly reject the serialized data if its
version is no longer supported, if the updated serialization format tracks the
serailzation version.

Some places that have limited versioning support:

`HloProgram` and `IfrtIRProgram` currently pin its internal serialization
including VHLO and IFRT IR to a at-least-4-week-old version. This may change to
accept a broader range of versions in case it is beneficial to be able to use
the latest SerDes version instead of 4-week-old version within a
single-versioned system.

`PluginProgram` ignores versioning. Any change to its format currently is a
breaking change because its serialization format is opaque/not extensible. Its
use will also be replaced with `CustomCallProgram`, so it seems not very
beneficial to add versioning to it.

IFRT Proxy does not yet use SerDes versioning. It will be done in a subsequent
change.

PiperOrigin-RevId: 759735059