Remove alias_info_ member from HloDataflowAnalysis.
It is only needed for calls to CanShareOperandBufferWithUser, and can be passed
by callers of this function. This makes it easier to avoid using default
parameters for AliasInfo, otherwise we would have to pass AliasInfo from
many users of HloDataflowAnalysis which do not actually use
CanShareOperandBufferWithUser.
This refactoring also shows that there were some users which used the default
parameter but actually should have passed correct AliasInfo. This will be
fixed in followup changes.

PiperOrigin-RevId: 775197668