Iterate through all fusion heroes when looking for a transpose emitter.
It's possible that a transpose fusion has several roots. Some of the roots will have transpose heroes that require going through shared memory. All these transpose heroes will have the same TransposeDescription and the same size of the element type. There might be other roots that are transposes, but they don't require shared memory and can be codegened by the loop emitter. We have to find the shmem transposes when deciding how to generate the code.

PiperOrigin-RevId: 775726340