Add some very slow algorithms to the cuDNN deny list
Engine 0 is sometimes really slow - in these cases it takes minutes instead of seconds. So I'm putting the relevant cases on the deny list.

I've also tried excluding engine 0 if other options are available but as it turns out there are cases where engine 0 is the better option over the more specialized other engines, so we can't do that.

PiperOrigin-RevId: 778872982