[XLA:GPU] fix scaled-dot emitter for handling 3d tensors with batch dimension.
The initial test set for scaled-dot missed to cover the test case with batch dims.
In this cl we fix that. We test that optimized hlo still has scaled-dot instruction and the final triton ir got dot_scaled instruction.

It:
a) drops the special handling of tile dims for scaled-dot op.
b) adds a new test case for 3D scaled-dot operations.
c) adds `ReshapeOp` and `TransOp` to `TritonXLAConvertUnsupportedTypesPass`.
PiperOrigin-RevId: 814614897