[PJRT C] Implement `Executable::GetOutputLayouts()` in the PJRT Layouts extension
This change implements a native support for `xla::Executable::GetOutputLayouts()` in PJRT C API, when PJRT Layouts extension is available. This support does not fetch the optimized HLO, and thus this method becomes more reliable and fast.

This change requires the plugin that implemented the Layouts extension
v2 to upgrade to v3 to avoid a incompatiblity.

PiperOrigin-RevId: 820819206