Enhance SymbolicExprContext and SymbolicMap Equality
We are introducing an operator== for SymbolicExprContext to allow comparison based on the underlying mlir::MLIRContext pointer and updating the SymbolicMap::operator== has been updated to use this new SymbolicExprContext::operator==.

This change ensures that two SymbolicMap instances are considered equal if they have the same number of dimensions, symbols, and expressions, even if they are associated with different SymbolicExprContext objects, as long as those contexts point to the same mlir::MLIRContext. This is crucial because SymbolicExpr instances are uniqued within a specific StorageUniquer, so expressions from contexts sharing the same MLIRContext are inherently equivalent. This is needed to make IndexingMapTest::IndexingMapSupportsAbslHashAndEqAndNe to pass with SymbolicMaps.

PiperOrigin-RevId: 828478294